chore: move constructLabelConfig out of label-config-editor as the parent component need to pass constructed labelConfig as props
